# openapi.yaml
openapi: 3.0.1
info:
  title: Document Management System API
  description: |
    A comprehensive Document Management System REST API with features including:

    - **Document Management**: Create, read, update, and delete documents
    - **Content Management**: Upload and download content with flexible storage options
    - **Versioning**: Major and minor version control for documents
    - **Content Renditions**: Primary and secondary content renditions with transformations
    - **Full-Text Search**: Lucene-powered search with fielded queries
    - **File Store Management**: Manage multiple file storage locations
    - **Content Transformations**: PDF-to-text and other content transformations

    ## Key Features

    ### Document Types
    - ARTICLE, MANUAL, REPORT, SPREADSHEET, PRESENTATION, IMAGE, VIDEO, AUDIO, OTHER

    ### Storage Options
    - Database storage for small files
    - File system storage for large files
    - Automatic migration between storage types

    ### Search Capabilities
    - Simple text search across all fields
    - Fielded search (name, description, keywords, tags, content)
    - Boolean operators (AND, OR)
    - Phrase search, wildcard search
    - Indexes all document metadata and indexable content

    ### Content Renditions
    - Primary renditions (original content)
    - Secondary renditions (transformed content)
    - Automatic text extraction from PDFs
    - Extensible transformer plugin framework
  contact:
...
  license:
...
servers:
...
paths:
...
  /api/search/rebuild:
    post:
      tags:
      - Search
      summary: Rebuild search index
      description: Completely rebuild the Lucene search index from all documents in
        the database. This indexes all document metadata and indexable content.
      operationId: rebuildIndex
      responses:
        "500":
          description: Failed to rebuild index
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Index rebuilt successfully
          content:
            '*/*':
              schema:
                type: string
  /api/search/fields:
    post:
      tags:
      - Search
      summary: Field-specific search
      description: Search specific fields with individual queries. Supports combining
        multiple field queries with AND or OR operators.
      operationId: searchFields
      parameters:
      - name: limit
        in: query
        description: Maximum number of results
        required: false
        schema:
          type: integer
          format: int32
          default: 50
        example: 50
      - name: operator
        in: query
        description: "Boolean operator: AND or OR"
        required: false
        schema:
          type: string
          default: AND
        example: AND
      requestBody:
        description: Map of field names to search queries
        content:
          application/json:
            schema:
              type: string
            examples:
              Multi-field search:
                description: Multi-field search
                value:
                  name: tutorial
                  tags: java
        required: true
      responses:
        "200":
          description: Search completed successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
        "400":
          description: Invalid query syntax
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
...
  /api/search:
    get:
      tags:
      - Search
      summary: Simple text search
      description: "Search across all document fields (name, description, keywords,\
        \ tags, content) and indexable content using Lucene query syntax. Supports\
        \ wildcards (*), phrase queries (\"exact phrase\"), and boolean operators\
        \ (AND, OR, NOT)."
      operationId: search
      parameters:
      - name: q
        in: query
        description: Search query using Lucene syntax
        required: true
        schema:
          type: string
        example: spring framework
      - name: limit
        in: query
        description: Maximum number of results to return
        required: false
        schema:
          type: integer
          format: int32
          default: 50
        example: 50
      responses:
        "200":
          description: Search completed successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
        "400":
          description: Invalid query syntax
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
  /api/search/stats:
    get:
      tags:
      - Search
      summary: Get search index statistics
      description: Retrieve statistics about the Lucene search index including document
        count and deleted documents.
      operationId: getStats
      responses:
        "200":
          description: Statistics retrieved successfully
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "500":
          description: Error retrieving statistics
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
...
components:
  schemas:
...
    SearchResult:
      type: object
      properties:
        documentId:
          type: integer
          format: int64
        score:
          type: number
          format: float
        name:
          type: string
        description:
          type: string
        keywords:
          type: string
        tags:
          type: string
...